Eliminated Bot::BasicBot in favor of my own bot code
Of course I borrowed a lot from Bot::BasicBot, but the primary chagne is the
support for multiple handlers, plus some events that Bot::BasicBot didn't
cover (from POE::Component::IRC::State). All tests pass, but I've killed off
the logging to the database. Towards getting that back, I'll next be:

* Adding tests for the new print logger, which is useful for debugging. Moving
* the database handle to its own module, so that it can be used from models or
  bot handlers. The main bot code won't even be aware of it anymore.
* Updating the config to support the new central DBI handle module.
* Updating the database schema, possibly adding new functions for modes and
  target users.
* Restoring the message logging code (now known as event logging) in
  App::Circle::Bot::Handler::Log.

I'm sure there's more, but this will do as a to-do list.